requirements.txt
django-3.1.6
pillow-8.2.0